Super basic privilage check around REST methods

+++ src/main/java/javaeetutorial/dukesbookstore/service/AbstractFacade.java
@@ -6,14 +6,23 @@
 package javaeetutorial.dukesbookstore.service;
 
 import java.util.List;
+import javaeetutorial.dukesbookstore.web.managedbeans.MemberSessionBean;
+import javax.enterprise.context.SessionScoped;
 import javax.persistence.EntityManager;
+import javax.xml.ws.http.HTTPException;
 
 /**
  *
  * @author Kyle.Lewer
  */
 public abstract class AbstractFacade<T> {
-
+    
+// use stateful memberSession to check if the user is a customer
+    // if the application required large scale and to be distributed across multiple services
+    // this would need to be transitioned to a stateless authentication method
+    @SessionScoped
+    MemberSessionBean memberSession;
+    
     private Class<T> entityClass;
 
     public AbstractFacade(Class<T> entityClass) {
@@ -21,44 +30,69 @@ public abstract class AbstractFacade<T> {
     }
 
     protected abstract EntityManager getEntityManager();
-
+    
     public void create(T entity) {
-        getEntityManager().persist(entity);
+        if(memberSession.isCustomer())
+            getEntityManager().persist(entity);
+        else
+            throw new HTTPException(401);
     }
 
     public void edit(T entity) {
-        getEntityManager().merge(entity);
+        if(memberSession.isCustomer())
+            getEntityManager().merge(entity);
+        else
+            throw new HTTPException(401);
     }
 
     public void remove(T entity) {
-        getEntityManager().remove(getEntityManager().merge(entity));
+        if(memberSession.isCustomer())
+            getEntityManager().remove(getEntityManager().merge(entity));
+        else
+            throw new HTTPException(401);
     }
 
     public T find(Object id) {
-        return getEntityManager().find(entityClass, id);
+        if(memberSession.isCustomer())
+            return getEntityManager().find(entityClass, id);
+        else
+            throw new HTTPException(401);
     }
 
     public List<T> findAll() {
-        javax.persistence.criteria.CriteriaQuery cq = getEntityManager().getCriteriaBuilder().createQuery();
-        cq.select(cq.from(entityClass));
-        return getEntityManager().createQuery(cq).getResultList();
+        if(memberSession.isCustomer()){
+            javax.persistence.criteria.CriteriaQuery cq = getEntityManager().getCriteriaBuilder().createQuery();
+            cq.select(cq.from(entityClass));
+            return getEntityManager().createQuery(cq).getResultList();
+        }else{
+            throw new HTTPException(401);
+        }
     }
 
     public List<T> findRange(int[] range) {
-        javax.persistence.criteria.CriteriaQuery cq = getEntityManager().getCriteriaBuilder().createQuery();
-        cq.select(cq.from(entityClass));
-        javax.persistence.Query q = getEntityManager().createQuery(cq);
-        q.setMaxResults(range[1] - range[0] + 1);
-        q.setFirstResult(range[0]);
-        return q.getResultList();
+        if(memberSession.isCustomer())
+        {
+            javax.persistence.criteria.CriteriaQuery cq = getEntityManager().getCriteriaBuilder().createQuery();
+            cq.select(cq.from(entityClass));
+            javax.persistence.Query q = getEntityManager().createQuery(cq);
+            q.setMaxResults(range[1] - range[0] + 1);
+            q.setFirstResult(range[0]);
+            return q.getResultList();
+        }else{
+            throw new HTTPException(401);
+        }
     }
 
     public int count() {
-        javax.persistence.criteria.CriteriaQuery cq = getEntityManager().getCriteriaBuilder().createQuery();
-        javax.persistence.criteria.Root<T> rt = cq.from(entityClass);
-        cq.select(getEntityManager().getCriteriaBuilder().count(rt));
-        javax.persistence.Query q = getEntityManager().createQuery(cq);
-        return ((Long) q.getSingleResult()).intValue();
+        if(memberSession.isCustomer()){
+            javax.persistence.criteria.CriteriaQuery cq = getEntityManager().getCriteriaBuilder().createQuery();
+            javax.persistence.criteria.Root<T> rt = cq.from(entityClass);
+            cq.select(getEntityManager().getCriteriaBuilder().count(rt));
+            javax.persistence.Query q = getEntityManager().createQuery(cq);
+            return ((Long) q.getSingleResult()).intValue();
+        } else {
+            throw new HTTPException(401);
+        }
     }
     
 }
